+++ src/models/trade.rs
@@ -4,5 +4,5 @@ pub struct Trade {
     pub buy_order_id: u64,
     pub sell_order_id: u64,
     pub quantity: u64,
-    pub price: f64,
+    pub price: u64,
 }
